Initial port

+++ src/common/literals.hpp
@@ -21,7 +21,7 @@ inline unsigned long long operator"" _TiB(unsigned long long x) {
 }
 
 // TODO: make sure the string's length doesn't exceed 8 characters
-inline constexpr u64 operator"" _u64(const char* str, unsigned long len) {
+inline constexpr u64 operator"" _u64(const char* str, unsigned long long len) {
     u64 res = 0;
     for (u32 i = 0; i < len; i++)
         res |= u64(str[i]) << (i * 8);
